task2final.py
This is the code for Task 2. I chose 2 different webiste's and wrote 2 different quotes for both the website's, as one of them did not have the 'doc-id', the 'rank' was chosen to be the unique id  to identify the applications, i.e the rank was considered as the 'doc-id' in this case. This is the 1st code for the 2nd task with the name task2final.py